# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CVImportProfile-Input": {
        "additionalProperties": false,
        "description": "Versioned, complete and immutable instructions for parsing one CSV shape.",
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnMapping"
            },
            "minItems": 1,
            "title": "Columns",
            "type": "array"
          },
          "decimal_convention": {
            "enum": [
              "point",
              "comma"
            ],
            "title": "Decimal Convention",
            "type": "string"
          },
          "delimiter": {
            "maxLength": 1,
            "minLength": 1,
            "title": "Delimiter",
            "type": "string"
          },
          "encoding": {
            "enum": [
              "utf-8",
              "utf-8-sig"
            ],
            "title": "Encoding",
            "type": "string"
          },
          "environment_id": {
            "minLength": 1,
            "title": "Environment Id",
            "type": "string"
          },
          "header_row": {
            "minimum": 1.0,
            "title": "Header Row",
            "type": "integer"
          },
          "metadata": {
            "$ref": "#/components/schemas/CVMetadata-Input"
          },
          "missing_value_tokens": {
            "items": {
              "type": "string"
            },
            "title": "Missing Value Tokens",
            "type": "array"
          },
          "schema_version": {
            "const": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "technique": {
            "const": "cyclic_voltammetry",
            "title": "Technique",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "technique",
          "environment_id",
          "encoding",
          "delimiter",
          "decimal_convention",
          "header_row",
          "missing_value_tokens",
          "columns",
          "metadata"
        ],
        "title": "CVImportProfile",
        "type": "object"
      },
      "CVImportProfile-Output": {
        "additionalProperties": false,
        "description": "Versioned, complete and immutable instructions for parsing one CSV shape.",
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnMapping"
            },
            "minItems": 1,
            "title": "Columns",
            "type": "array"
          },
          "decimal_convention": {
            "enum": [
              "point",
              "comma"
            ],
            "title": "Decimal Convention",
            "type": "string"
          },
          "delimiter": {
            "maxLength": 1,
            "minLength": 1,
            "title": "Delimiter",
            "type": "string"
          },
          "encoding": {
            "enum": [
              "utf-8",
              "utf-8-sig"
            ],
            "title": "Encoding",
            "type": "string"
          },
          "environment_id": {
            "minLength": 1,
            "title": "Environment Id",
            "type": "string"
          },
          "header_row": {
            "minimum": 1.0,
            "title": "Header Row",
            "type": "integer"
          },
          "metadata": {
            "$ref": "#/components/schemas/CVMetadata-Output"
          },
          "missing_value_tokens": {
            "items": {
              "type": "string"
            },
            "title": "Missing Value Tokens",
            "type": "array"
          },
          "schema_version": {
            "const": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "technique": {
            "const": "cyclic_voltammetry",
            "title": "Technique",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "technique",
          "environment_id",
          "encoding",
          "delimiter",
          "decimal_convention",
          "header_row",
          "missing_value_tokens",
          "columns",
          "metadata"
        ],
        "title": "CVImportProfile",
        "type": "object"
      },
      "CVLineage": {
        "additionalProperties": false,
        "properties": {
          "environment_id": {
            "minLength": 1,
            "title": "Environment Id",
            "type": "string"
          },
          "import_profile_id": {
            "minLength": 1,
            "title": "Import Profile Id",
            "type": "string"
          },
          "source_artifact_id": {
            "minLength": 1,
            "title": "Source Artifact Id",
            "type": "string"
          },
          "source_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Source Sha256",
            "type": "string"
          },
          "transformation_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Transformation Ids",
            "type": "array"
          }
        },
        "required": [
          "environment_id",
          "source_artifact_id",
          "source_sha256",
          "import_profile_id",
          "transformation_ids"
        ],
        "title": "CVLineage",
        "type": "object"
      },
      "CVMetadata-Input": {
        "additionalProperties": false,
        "description": "Phase 2 states required to avoid inventing electrochemical context.",
        "properties": {
          "contact_area": {
            "$ref": "#/components/schemas/MetadataValue-Input"
          },
          "current_basis": {
            "$ref": "#/components/schemas/MetadataValue-Input"
          },
          "cycle_information": {
            "$ref": "#/components/schemas/MetadataValue-Input"
          },
          "electrode_role": {
            "$ref": "#/components/schemas/MetadataValue-Input"
          },
          "geometric_area": {
            "$ref": "#/components/schemas/MetadataValue-Input"
          },
          "potential_treatment": {
            "$ref": "#/components/schemas/MetadataValue-Input"
          },
          "reference_scale": {
            "$ref": "#/components/schemas/MetadataValue-Input"
          },
          "scan_rate": {
            "$ref": "#/components/schemas/MetadataValue-Input"
          }
        },
        "required": [
          "reference_scale",
          "potential_treatment",
          "current_basis",
          "electrode_role",
          "geometric_area",
          "contact_area",
          "scan_rate",
          "cycle_information"
        ],
        "title": "CVMetadata",
        "type": "object"
      },
      "CVMetadata-Output": {
        "additionalProperties": false,
        "description": "Phase 2 states required to avoid inventing electrochemical context.",
        "properties": {
          "contact_area": {
            "$ref": "#/components/schemas/MetadataValue-Output"
          },
          "current_basis": {
            "$ref": "#/components/schemas/MetadataValue-Output"
          },
          "cycle_information": {
            "$ref": "#/components/schemas/MetadataValue-Output"
          },
          "electrode_role": {
            "$ref": "#/components/schemas/MetadataValue-Output"
          },
          "geometric_area": {
            "$ref": "#/components/schemas/MetadataValue-Output"
          },
          "potential_treatment": {
            "$ref": "#/components/schemas/MetadataValue-Output"
          },
          "reference_scale": {
            "$ref": "#/components/schemas/MetadataValue-Output"
          },
          "scan_rate": {
            "$ref": "#/components/schemas/MetadataValue-Output"
          }
        },
        "required": [
          "reference_scale",
          "potential_treatment",
          "current_basis",
          "electrode_role",
          "geometric_area",
          "contact_area",
          "scan_rate",
          "cycle_information"
        ],
        "title": "CVMetadata",
        "type": "object"
      },
...
      "ColumnMapping": {
        "additionalProperties": false,
        "description": "One explicit decision about one source column.",
        "properties": {
          "role": {
            "enum": [
              "potential",
              "current",
              "current_density",
              "time",
              "cycle",
              "ignored"
            ],
            "title": "Role",
            "type": "string"
          },
          "source_column": {
            "minLength": 1,
            "title": "Source Column",
            "type": "string"
          },
          "source_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Unit"
          },
          "target_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Unit"
          }
        },
        "required": [
          "source_column",
          "role"
        ],
        "title": "ColumnMapping",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InspectionRequest": {
        "additionalProperties": false,
        "properties": {
          "delimiter": {
            "maxLength": 1,
            "minLength": 1,
            "title": "Delimiter",
            "type": "string"
          },
          "encoding": {
            "enum": [
              "utf-8",
              "utf-8-sig"
            ],
            "title": "Encoding",
            "type": "string"
          },
          "header_row": {
            "minimum": 1.0,
            "title": "Header Row",
            "type": "integer"
          },
          "source_artifact_id": {
            "title": "Source Artifact Id",
            "type": "string"
          }
        },
        "required": [
          "encoding",
          "delimiter",
          "header_row",
          "source_artifact_id"
        ],
        "title": "InspectionRequest",
        "type": "object"
      },
...
      "MetadataValue-Input": {
        "additionalProperties": false,
        "description": "One metadata value whose evidence state is independent from its value.",
        "properties": {
          "state": {
            "enum": [
              "known",
              "unknown",
              "unavailable",
              "not_applicable"
            ],
            "title": "State",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "state"
        ],
        "title": "MetadataValue",
        "type": "object"
      },
      "MetadataValue-Output": {
        "additionalProperties": false,
        "description": "One metadata value whose evidence state is independent from its value.",
        "properties": {
          "state": {
            "enum": [
              "known",
              "unknown",
              "unavailable",
              "not_applicable"
            ],
            "title": "State",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "state"
        ],
        "title": "MetadataValue",
        "type": "object"
      },
      "NormalisationRequest": {
        "additionalProperties": false,
        "properties": {
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "source_artifact_id": {
            "title": "Source Artifact Id",
            "type": "string"
          }
        },
        "required": [
          "source_artifact_id",
          "profile_id"
        ],
        "title": "NormalisationRequest",
        "type": "object"
      },
      "NormalisationResult": {
        "additionalProperties": false,
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/StructuralFinding"
            },
            "title": "Findings",
            "type": "array"
          },
          "graph": {
            "$ref": "#/components/schemas/TransformationGraph"
          },
          "observation": {
            "$ref": "#/components/schemas/NormalisedCVObservation"
          }
        },
        "required": [
          "observation",
          "graph",
          "findings"
        ],
        "title": "NormalisationResult",
        "type": "object"
      },
      "NormalisationView": {
        "properties": {
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "result": {
            "$ref": "#/components/schemas/NormalisationResult"
          }
        },
        "required": [
          "result",
          "replayed"
        ],
        "title": "NormalisationView",
        "type": "object"
      },
...
      "NormalisedSeries": {
        "additionalProperties": false,
        "properties": {
          "dtype": {
            "const": "decimal",
            "title": "Dtype",
            "type": "string"
          },
          "role": {
            "enum": [
              "potential",
              "current",
              "current_density",
              "time",
              "cycle",
              "ignored"
            ],
            "title": "Role",
            "type": "string"
          },
          "schema_version": {
            "const": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "series_id": {
            "minLength": 1,
            "title": "Series Id",
            "type": "string"
          },
          "shape": {
            "maxItems": 1,
            "minItems": 1,
            "prefixItems": [
              {
                "type": "integer"
              }
            ],
            "title": "Shape",
            "type": "array"
          },
          "source_column": {
            "minLength": 1,
            "title": "Source Column",
            "type": "string"
          },
          "source_unit": {
            "minLength": 1,
            "title": "Source Unit",
            "type": "string"
          },
          "transformation_id": {
            "minLength": 1,
            "title": "Transformation Id",
            "type": "string"
          },
          "unit": {
            "minLength": 1,
            "title": "Unit",
            "type": "string"
          },
          "values": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "series_id",
          "schema_version",
          "dtype",
          "shape",
          "source_column",
          "role",
          "source_unit",
          "unit",
          "values",
          "transformation_id"
        ],
        "title": "NormalisedSeries",
        "type": "object"
      },
...
      "PlotSeriesView": {
        "properties": {
          "data_origin": {
            "title": "Data Origin",
            "type": "string"
          },
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "execution_mode": {
            "title": "Execution Mode",
            "type": "string"
          },
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/CVLineage"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/NormalisedSeries"
            },
            "title": "Series",
            "type": "array"
          }
        },
        "required": [
          "observation_id",
          "data_origin",
          "execution_mode",
          "environment_id",
          "series",
          "provenance"
        ],
        "title": "PlotSeriesView",
        "type": "object"
      },
      "ProfileView": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/CVImportProfile-Output"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          }
        },
        "required": [
          "profile_id",
          "profile",
          "replayed"
        ],
        "title": "ProfileView",
        "type": "object"
      },
...
      "SourceInspectionView": {
        "properties": {
          "headers": {
            "items": {
              "type": "string"
            },
            "title": "Headers",
            "type": "array"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "source_artifact_id": {
            "title": "Source Artifact Id",
            "type": "string"
          },
          "source_sha256": {
            "title": "Source Sha256",
            "type": "string"
          }
        },
        "required": [
          "source_artifact_id",
          "source_sha256",
          "headers",
          "row_count"
        ],
        "title": "SourceInspectionView",
        "type": "object"
      },
      "StructuralFinding": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "finding_id": {
            "minLength": 1,
            "title": "Finding Id",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "status": {
            "enum": [
              "pass",
              "warning",
              "blocking"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject_id": {
            "minLength": 1,
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "finding_id",
          "code",
          "status",
          "subject_id",
          "message"
        ],
        "title": "StructuralFinding",
        "type": "object"
      },
      "TransformationGraph": {
        "additionalProperties": false,
        "properties": {
          "observation_id": {
            "minLength": 1,
            "title": "Observation Id",
            "type": "string"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/TransformationRecord"
            },
            "minItems": 1,
            "title": "Records",
            "type": "array"
          },
          "source_artifact_id": {
            "minLength": 1,
            "title": "Source Artifact Id",
            "type": "string"
          }
        },
        "required": [
          "source_artifact_id",
          "observation_id",
          "records"
        ],
        "title": "TransformationGraph",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LabBridge",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cv/import-profiles": {
      "post": {
        "operationId": "create_profile_cv_import_profiles_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CVImportProfile-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Profile"
      }
    },
    "/cv/import-profiles/{profile_id}": {
      "get": {
        "operationId": "read_profile_cv_import_profiles__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Profile"
      }
    },
    "/cv/normalisations": {
      "post": {
        "operationId": "normalise_cv_normalisations_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NormalisationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalisationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Normalise"
      }
    },
    "/cv/normalised-observations/{observation_id}": {
      "get": {
        "operationId": "read_normalisation_cv_normalised_observations__observation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalisationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Normalisation"
      }
    },
    "/cv/normalised-observations/{observation_id}/plot-series": {
      "get": {
        "operationId": "plot_series_cv_normalised_observations__observation_id__plot_series_get",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotSeriesView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plot Series"
      }
    },
    "/cv/source-inspections": {
      "post": {
        "operationId": "inspect_source_cv_source_inspections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceInspectionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect Source"
      }
    },
...
  }
}